+++ tests/JobTest.php
@@ -15,7 +15,7 @@ class JobTest extends QlessTestCase
     /**
      * @test
      * @expectedException \Qless\Exceptions\RuntimeException
-     * @expectedExceptionMessage Could not find job class Sample\TestWorkerImpl.
+     * @expectedExceptionMessage Could not find job class MyJobClass.
      */
     public function shouldThrowsExpectedExceptionWhenGetInstanceWithNonExistentClass()
     {
@@ -24,15 +24,12 @@ class JobTest extends QlessTestCase
         $this->client->config->set('heartbeat', -10);
         $this->client->config->set('grace-period', 0);
 
-        $queue->put(
-            'Sample\TestWorkerImpl',
-            ['performMethod' => 'myPerformMethod', 'payload' => 'otherData']
-        );
+        $queue->put('MyJobClass', ['performMethod' => 'myPerformMethod', 'payload' => 'otherData']);
 
-        $job1 = $queue->pop('worker-1')[0];
-        $queue->pop('worker-2');
+        $job = $queue->pop();
+        $queue->pop();
 
-        $job1->getInstance();
+        $job->getInstance();
     }
 
     /**
@@ -47,15 +44,12 @@ class JobTest extends QlessTestCase
         $this->client->config->set('heartbeat', -10);
         $this->client->config->set('grace-period', 0);
 
-        $queue->put(
-            \stdClass::class,
-            ['performMethod' => 'myPerformMethod', 'payload' => 'otherData']
-        );
+        $queue->put('stdClass', ['performMethod' => 'myPerformMethod', 'payload' => 'otherData']);
 
-        $job1 = $queue->pop('worker-1')[0];
-        $queue->pop('worker-2');
+        $job = $queue->pop();
+        $queue->pop();
 
-        $job1->getInstance();
+        $job->getInstance();
     }
 
     /**
@@ -73,15 +67,12 @@ class JobTest extends QlessTestCase
         $this->client->config->set('heartbeat', -10);
         $this->client->config->set('grace-period', 0);
 
-        $queue->put(
-            WorkerStub2::class,
-            ['performMethod' => 'myPerformMethod2', 'payload' => 'otherData']
-        );
+        $queue->put(WorkerStub2::class, ['performMethod' => 'myPerformMethod2', 'payload' => 'otherData']);
 
-        $job1 = $queue->pop('worker-1')[0];
-        $queue->pop('worker-2');
+        $job = $queue->pop();
+        $queue->pop();
 
-        $job1->getInstance();
+        $job->getInstance();
     }
 
     /** @test */
@@ -92,15 +83,12 @@ class JobTest extends QlessTestCase
         $this->client->config->set('heartbeat', -10);
         $this->client->config->set('grace-period', 0);
 
-        $queue->put(
-            WorkerStub2::class,
-            ['performMethod' => 'myPerformMethod', 'payload' => 'otherData']
-        );
+        $queue->put(WorkerStub2::class, ['performMethod' => 'myPerformMethod', 'payload' => 'otherData']);
 
-        $job1 = $queue->pop('worker-1')[0];
-        $queue->pop('worker-2');
+        $job = $queue->pop();
+        $queue->pop();
 
-        $this->assertInstanceOf(WorkerStub2::class, $job1->getInstance());
+        $this->assertInstanceOf(WorkerStub2::class, $job->getInstance());
     }
 
     /** @test */
@@ -113,10 +101,10 @@ class JobTest extends QlessTestCase
 
         $queue->put(WorkerStub2::class, ['payload' => 'otherData']);
 
-        $job1 = $queue->pop('worker-1')[0];
-        $queue->pop('worker-2');
+        $job = $queue->pop();
+        $queue->pop();
 
-        $this->assertInstanceOf(WorkerStub2::class, $job1->getInstance());
+        $this->assertInstanceOf(WorkerStub2::class, $job->getInstance());
     }
 
     /**
@@ -131,23 +119,21 @@ class JobTest extends QlessTestCase
         $this->client->config->set('heartbeat', -10);
         $this->client->config->set('grace-period', 0);
 
-        $queue->put(
-            'Sample\TestWorkerImpl',
-            ['performMethod' => 'myPerformMethod', 'payload' => 'otherData']
-        );
+        $queue->put('MyJobClass', ['performMethod' => 'myPerformMethod', 'payload' => 'otherData']);
 
-        $job1 = $queue->pop('worker-1')[0];
+        $job = $queue->pop('worker-1');
         $queue->pop('worker-2');
-        $job1->heartbeat();
+
+        $job->heartbeat();
     }
 
-    public function testCanGetCorrectTTL()
+    /** @test */
+    public function shouldGetCorrectTtl()
     {
         $queue = new Queue('testQueue', $this->client);
-        $queue->put("Sample\\TestWorkerImpl", []);
-        $job = $queue->pop('worker-1')[0];
-        $ttl = $job->ttl();
-        $this->assertGreaterThan(55, $ttl);
+        $queue->put('MyJobClass', []);
+
+        $this->assertGreaterThan(55, $queue->pop()->ttl());
     }
 
     public function testCompleteJob()
@@ -157,24 +143,18 @@ class JobTest extends QlessTestCase
         $testData = ["performMethod" => 'myPerformMethod', "payload" => "otherData"];
         $queue->put("Sample\\TestWorkerImpl", $testData);
 
-        $job1 = $queue->pop('worker-1')[0];
-        $res = $job1->complete();
-        $this->assertEquals('complete', $res);
+        $this->assertEquals('complete', $queue->pop()->complete());
     }
 
-    public function testFailJobCannotBePopped()
+    /** @test */
+    public function shouldNotPopFailedJob()
     {
         $queue = new Queue('testQueue', $this->client);
 
-        $testData = ["performMethod" => 'myPerformMethod', "payload" => "otherData"];
-        $queue->put("Sample\\TestWorkerImpl", $testData, 'jid');
+        $queue->put('MyJobClass', [], 'jid');
 
-        $job1 = $queue->pop('worker-1')[0];
-        $res = $job1->fail('account', 'failed to connect');
-        $this->assertEquals('jid', $res);
-
-        $job1 = $queue->pop('worker-1');
-        $this->assertEmpty($job1);
+        $this->assertEquals('jid', $queue->pop()->fail('account', 'failed to connect'));
+        $this->assertNull($queue->pop('worker-1'));
     }
 
     public function testRetryDoesReturnJobAndDefaultsToFiveRetries()
@@ -182,13 +162,13 @@ class JobTest extends QlessTestCase
         $queue = new Queue('testQueue', $this->client);
 
         $testData = ["performMethod" => 'myPerformMethod', "payload" => "otherData"];
-        $queue->put('Sample\TestWorkerImpl', $testData, 'jid');
+        $queue->put('MyJobClass', $testData, 'jid');
 
-        $job1 = $queue->pop('worker-1')[0];
+        $job1 = $queue->pop();
         $remaining = $job1->retry('account', 'failed to connect');
         $this->assertEquals(4, $remaining);
 
-        $job1 = $queue->pop('worker-1')[0];
+        $job1 = $queue->pop();
         $this->assertEquals('jid', $job1->getId());
     }
 
@@ -196,15 +176,10 @@ class JobTest extends QlessTestCase
     {
         $queue = new Queue('testQueue', $this->client);
 
-        $testData = ["performMethod" => 'myPerformMethod', "payload" => "otherData"];
-        $queue->put('Sample\TestWorkerImpl', $testData, 'jid', 0, 1);
+        $queue->put('MyJobClass', [], 'jid', 0, 1);
 
-        $job1 = $queue->pop('worker-1')[0];
-        $remaining = $job1->retry('account', 'failed to connect');
-        $this->assertEquals(0, $remaining);
-
-        $job1 = $queue->pop('worker-1')[0];
-        $this->assertEquals('jid', $job1->getId());
+        $this->assertZero($queue->pop()->retry('account', 'failed to connect'));
+        $this->assertEquals('jid', $queue->pop()->getId());
     }
 
     public function testRetryDoesReturnNegativeWhenNoMoreAvailable()
@@ -212,9 +187,9 @@ class JobTest extends QlessTestCase
         $queue = new Queue('testQueue', $this->client);
 
         $testData = ["performMethod" => 'myPerformMethod', "payload" => "otherData"];
-        $queue->put('Sample\TestWorkerImpl', $testData, 'jid', 0, 0);
+        $queue->put('MyJobClass', $testData, 'jid', 0, 0);
 
-        $job1 = $queue->pop('worker-1')[0];
+        $job1 = $queue->pop();
         $remaining = $job1->retry('account', 'failed to connect');
         $this->assertEquals(-1, $remaining);
     }
@@ -224,41 +199,38 @@ class JobTest extends QlessTestCase
         $queue = new Queue('testQueue', $this->client);
 
         $testData = ["performMethod" => 'myPerformMethod', "payload" => "otherData"];
-        $queue->put('Sample\TestWorkerImpl', $testData, 'jid', 0, 0);
+        $queue->put('MyJobClass', $testData, 'jid', 0, 0);
 
-        $job1 = $queue->pop('worker-1')[0];
-        $job1->retry('account', 'failed to connect');
+        $job = $queue->pop();
+        $job->retry('account', 'failed to connect');
 
-        $job1 = $queue->pop('worker-1');
-        $this->assertEmpty($job1);
+        $this->assertNull($queue->pop());
     }
 
-    public function testCancelRemovesJob()
+    /** @test */
+    public function shouldCancelRemovesJob()
     {
         $queue = new Queue('testQueue', $this->client);
 
         $data = ['performMethod' => 'myPerformMethod', 'payload' => 'otherData'];
-        $queue->put('Sample\TestWorkerImpl', $data, 'jid-1', 0, 0);
-        $queue->put('Sample\TestWorkerImpl', $data, 'jid-2', 0, 0);
 
-        $job1 = $queue->pop('worker-1')[0];
-        $res = $job1->cancel();
+        $queue->put('MyJobClass', $data, 'jid-1', 0, 0);
+        $queue->put('MyJobClass', $data, 'jid-2', 0, 0);
 
-        $this->assertEquals(['jid-1'], $res);
+        $this->assertEquals(['jid-1'], $queue->pop()->cancel());
     }
 
-    public function testCancelRemovesJobWithDependents()
+    /** @test */
+    public function shouldCancelRemovesJobWithDependents()
     {
         $queue = new Queue('testQueue', $this->client);
 
         $data = ['performMethod' => 'myPerformMethod', 'payload' => 'otherData'];
-        $queue->put('Sample\TestWorkerImpl', $data, 'jid-1', 0, 0);
-        $queue->put('Sample\TestWorkerImpl', $data, 'jid-2', 0, 0, true, 0, [], 0, [], ['jid-1']);
 
-        $job1 = $queue->pop('worker-1')[0];
-        $res = $job1->cancel(true);
+        $queue->put('MyJobClass', $data, 'jid-1', 0, 0);
+        $queue->put('MyJobClass', $data, 'jid-2', 0, 0, true, 0, [], 0, [], ['jid-1']);
 
-        $this->assertEquals(['jid-1', 'jid-2'], $res);
+        $this->assertEquals(['jid-1', 'jid-2'], $queue->pop()->cancel(true));
     }
 
     /**
@@ -269,22 +241,19 @@ class JobTest extends QlessTestCase
         $queue = new Queue('testQueue', $this->client);
 
         $data = ['performMethod' => 'myPerformMethod', 'payload' => 'otherData'];
-        $queue->put('Sample\TestWorkerImpl', $data, 'jid-1', 0, 0);
-        $queue->put('Sample\TestWorkerImpl', $data, 'jid-2', 0, 0, true, 0, [], 0, [], ['jid-1']);
 
-        $job1 = $queue->pop('worker-1')[0];
-        $job1->cancel();
+        $queue->put('MyJobClass', $data, 'jid-1', 0, 0);
+        $queue->put('MyJobClass', $data, 'jid-2', 0, 0, true, 0, [], 0, [], ['jid-1']);
+
+        $queue->pop()->cancel();
     }
 
     public function testItCanAddTagsToAJobWithNoExistingTags()
     {
         $queue = new Queue('testQueue', $this->client);
 
-        $data = ['performMethod' => 'myPerformMethod', 'payload' => 'otherData'];
-        $queue->put('Sample\TestWorkerImpl', $data, 'jid-1', 0, 0);
-
-        $job1 = $queue->pop('worker-1')[0];
-        $job1->tag('a', 'b');
+        $queue->put('MyJobClass', [], 'jid-1', 0, 0);
+        $queue->pop()->tag('a', 'b');
 
         $data = json_decode($this->client->get('jid-1'));
         $this->assertEquals(['a', 'b'], $data->tags);
@@ -294,11 +263,8 @@ class JobTest extends QlessTestCase
     {
         $queue = new Queue('testQueue', $this->client);
 
-        $data = ['performMethod' => 'myPerformMethod', 'payload' => 'otherData'];
-        $queue->put('Sample\TestWorkerImpl', $data, 'jid-1', 0, 0, true, 0, [], 0, ['1', '2']);
-
-        $job1 = $queue->pop('worker-1')[0];
-        $job1->tag('a', 'b');
+        $queue->put('MyJobClass', [], 'jid-1', 0, 0, true, 0, [], 0, ['1', '2']);
+        $queue->pop()->tag('a', 'b');
 
         $data = json_decode($this->client->get('jid-1'));
         $this->assertEquals(['1', '2', 'a', 'b'], $data->tags);
@@ -308,11 +274,8 @@ class JobTest extends QlessTestCase
     {
         $queue = new Queue('testQueue', $this->client);
 
-        $data = ['performMethod' => 'myPerformMethod', 'payload' => 'otherData'];
-        $queue->put('Sample\TestWorkerImpl', $data, 'jid-1', 0, 0, true, 0, [], 0, ['1', '2', '3']);
-
-        $job1 = $queue->pop('worker-1')[0];
-        $job1->untag('2', '3');
+        $queue->put('MyJobClass', [], 'jid-1', 0, 0, true, 0, [], 0, ['1', '2', '3']);
+        $queue->pop()->untag('2', '3');
 
         $data = json_decode($this->client->get('jid-1'));
         $this->assertEquals(['1'], $data->tags);
@@ -322,13 +285,11 @@ class JobTest extends QlessTestCase
     {
         $queue = new Queue('testQueue', $this->client);
 
-        $data = ['performMethod' => 'myPerformMethod', 'payload' => 'otherData'];
-        $queue->put('Sample\TestWorkerImpl', $data, 'jid-1', 0, 0, true, 1, [], 5, ['tag1','tag2']);
+        $queue->put('MyJobClass', [], 'jid-1', 0, 0, true, 1, [], 5, ['tag1','tag2']);
+        $queue->pop()->requeue();
 
-        $job = $queue->pop('worker-1')[0];
-        $job->requeue();
+        $job = $queue->pop();
 
-        $job = $queue->pop('worker-1')[0];
         $this->assertEquals(1, $job->getPriority());
         $this->assertEquals(['tag1','tag2'], $job->getTags());
     }
@@ -337,13 +298,10 @@ class JobTest extends QlessTestCase
     {
         $queue = new Queue('testQueue', $this->client);
 
-        $data = ['performMethod' => 'myPerformMethod', 'payload' => 'otherData'];
-        $queue->put('Sample\TestWorkerImpl', $data, 'jid-1', 0, 0, true, 1, [], 5, ['tag1','tag2']);
+        $queue->put('MyJobClass', [], 'jid-1', 0, 0, true, 1, [], 5, ['tag1','tag2']);
+        $queue->pop()->requeue(['tags' => ['nnn']]);
 
-        $job = $queue->pop('worker-1')[0];
-        $job->requeue(['tags' => ['nnn']]);
-
-        $job = $queue->pop('worker-1')[0];
+        $job = $queue->pop();
         $this->assertEquals(1, $job->getPriority());
         $this->assertEquals(['nnn'], $job->getTags());
     }
@@ -356,9 +314,9 @@ class JobTest extends QlessTestCase
         $queue = new Queue('testQueue', $this->client);
 
         $data = ['performMethod' => 'myPerformMethod', 'payload' => 'otherData'];
-        $queue->put('Sample\TestWorkerImpl', $data, 'jid-1', 0, 0, true, 1, [], 5, ['tag1','tag2']);
+        $queue->put('MyJobClass', $data, 'jid-1', 0, 0, true, 1, [], 5, ['tag1','tag2']);
 
-        $job = $queue->pop('worker-1')[0];
+        $job = $queue->pop();
         $this->client->cancel('jid-1');
         $job->requeue();
     }